Only perform exact match when [in|ex]cluding directories
The flags `--terragrunt-include-dir` and `--terragrunt-exclude-dir` are
“Unix-style glob of directories to include (resp. exclude) when running
*-all commands”

Currently, the matching on which directories to include/exclude is done
using `string.Contains`, which means that this excluding `module-a` would
also exclude `module-abba`.

We should really be doing an exact match, since we can get the multi-match
behaviour by using Unix-style globs.

Also add two unit-tests confirming the old and new behaviours.